Remove unusued imports and improve some code style.

+++ src/main/java/ivorius/reccomplex/items/ItemBlockGenericSpace.java
@@ -40,7 +40,7 @@ public class ItemBlockGenericSpace extends ItemBlock
         IBlockState iblockstate = worldIn.getBlockState(pos);
         Block block = iblockstate.getBlock();
 
-        if (block == Blocks.snow_layer && ((Integer) iblockstate.getValue(BlockSnow.LAYERS)).intValue() < 1)
+        if (block == Blocks.snow_layer && iblockstate.getValue(BlockSnow.LAYERS) < 1)
         {
             side = EnumFacing.UP;
         }
